requirements.txt
Django==1.10.2
django-braces==1.10.0
psycopg2==2.6.2
